refactor: move format methods to utils
