+++ src/components/staking/HistoryTab.vue
@@ -1,44 +1,21 @@
 <script lang="ts">
 import { defineComponent, computed } from 'vue';
 import { useStore } from 'src/store';
-import { Action, RexHistory } from 'src/types';
-import { getChain } from 'src/config/ConfigManager';
-import { assetToAmount } from 'src/utils/string-utils';
+import { Action } from 'src/types';
+import { getRexHistoryAsset, formatDate } from 'src/utils/string-utils';
 
 export default defineComponent({
     name: 'HistoryTab',
     components: {},
     setup() {
         const store = useStore();
-        const symbol = getChain().getSystemToken().symbol;
         const rexActions = computed((): Action[] => store.state.account.rexActions);
-
-        function formatDate(date: string): string {
-            return new Date(date).toLocaleDateString('en-US', {
-                month: 'long',
-                year: 'numeric',
-                day: 'numeric',
-            });
-        }
-
-        function getHistoryAmount(data: RexHistory): string {
-            if (data.rex){
-                return data.rex;
-            }
-            if (typeof data.amount === 'number') {
-                const total = (assetToAmount(data.from_cpu) + assetToAmount(data.from_net));
-                return `${total} ${symbol}`;
-            }else{
-                return data.amount;
-            }
-        }
 
         return {
             store,
-            symbol,
             rexActions,
             formatDate,
-            getHistoryAmount,
+            getRexHistoryAsset,
         };
     },
 });
@@ -56,7 +33,7 @@ export default defineComponent({
             <div class="col-xs-12 col-sm-6">
                 <div class="row q-pa-sm">
                     <div class="col-6">{{action.act.name}}</div>
-                    <div class="col-6 text-weight-bold">{{ getHistoryAmount(action.act.data)}}</div>
+                    <div class="col-6 text-weight-bold">{{ getRexHistoryAsset(action.act.data)}}</div>
                 </div>
             </div>
             <div class="col-xs-12 col-sm-6">
